scripts/offline_runtime_requirements.txt
# Extra runtime deps needed in addition to root requirements.txt + torch/torchvision
# for running Neural-SLAM main loop with Habitat stack.

gym==0.10.9
yacs>=0.1.5
numpy-quaternion>=2019.3.18.14.33.20
opencv-python>=3.3.0
tqdm>=4.0.0
numba>=0.44.0
attrs>=19.1.0
imageio-ffmpeg>=0.2.0
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scikit-fmm==2019.1.30
